// src/app/models/listPeople.model.ts
import { StarWars } from './starWars.model';
import { Veiculos } from './veiculos.model';

export interface Result {
  name?: string;
  height?: string;
  mass?: string;
  hair_color?: string;
  skin_color?: string;
  eye_color?: string;
  birth_year?: string;
  gender?: string;
  homeworld?: string;
  films?: string[];
  species?: string[];
  vehicles?:Veiculos[];
  starships?: string[];
  created?: Date;
  edited?: Date;
  url?: string;
  people?:StarWars;
}

export interface ListPeople {
  count?: number;
  next?: string;
  previous?: any;
  results?: Result[];
}